IMPROVEMENT: Rename ResourceManipulator to ThumbnailManipulator
