+++ DistributionPackages/Neos.Demo.UserRegistration/Classes/Package.php
@@ -2,7 +2,7 @@
 namespace Neos\Demo\UserRegistration;
 
 use Neos\Demo\UserRegistration\Domain\Service\Asset\AssetManipulator;
-use Neos\Demo\UserRegistration\Domain\Service\Resource\ResourceManipulator;
+use Neos\Demo\UserRegistration\Domain\Service\Asset\ThumbnailManipulator;
 use Neos\Flow\Core\Bootstrap;
 use Neos\Flow\Package\Package as BasePackage;
 use Neos\Media\Domain\Service\ThumbnailService;
@@ -16,6 +16,6 @@ class Package extends BasePackage
         $dispatcher = $bootstrap->getSignalSlotDispatcher();
         $dispatcher->connect(AssetService::class, 'assetCreated', AssetManipulator::class, 'assetCreated');
         $dispatcher->connect(ContentController::class, 'assetUploaded', AssetManipulator::class, 'assetUploaded');
-        $dispatcher->connect(ThumbnailService::class, 'thumbnailCreated', ResourceManipulator::class, 'thumbnailCreated');
+        $dispatcher->connect(ThumbnailService::class, 'thumbnailCreated', ThumbnailManipulator::class, 'thumbnailCreated');
     }
 }
